Change(commands): overall improvement
* (chore) prevent removing logs directory
* `study` => `bgm`
* `dc` => `disconnect`
* add an alias `support` to `help`
* `mv` => `move`
* `np` => `nowplaying`
* split into `update` and `ping`
* `rmall` => `removeall`
