+++ docs/docs/guide/commands/bgm.md
@@ -1,14 +1,15 @@
 ---
-sidebar_label: bgm
+sidebar_label: BGM
 ---
-# `bgm`コマンド
+# `BGM`コマンド
 開発者が勝手に作った勉強用・作業用BGMのプリセットプレイリストを表示し、聞きたいものを選択して再生することができます。
 
-スラッシュコマンドでは、`/study`を使用してください。
+スラッシュコマンドでは、`/bgm`を使用してください。
 
 ## 別名
-`bgm`以外にも以下の別名を使用できます。
+`BGM`以外にも以下の別名を使用できます。
 
+- bgm
 - study
