+++ docs/docs/guide/commands/disconnect.md
@@ -4,14 +4,14 @@ sidebar_label: 切断
 # `切断`コマンド
 ボイスチャンネルから切断します。
 
-スラッシュコマンドでは、`/dc`を使用してください。
+スラッシュコマンドでは、`/disconnect`を使用してください。
 
 ## 別名
 `切断`以外にも以下の別名を使用できます。
 
 - 終了
-- dc
 - disconnect
+- dc
 - leave
 - quit
