+++ docs/docs/guide/commands/move.md
@@ -4,13 +4,13 @@ sidebar_label: 移動
 # `移動`コマンド
 曲を指定された位置から指定された位置までキュー内で移動します。2番目の曲を5番目に移動したい場合は`mv 2 5`のようにします。
 
-スラッシュコマンドでは、`/mv`を使用してください。
+スラッシュコマンドでは、`/move`を使用してください。
 
 ## 別名
 `移動`以外にも以下の別名を使用できます。
 
-- mv
 - move
+- mv
 
 ## 使い方
 ```
