+++ docs/docs/guide/commands/nowplaying.md
@@ -4,14 +4,14 @@ sidebar_label: 現在再生中
 # `現在再生中`コマンド
 現在再生中の曲の情報を表示します。 `l`(スラッシュコマンドの場合はTrue)を引数にするとより長く概要を表示します(可能な場合)
 
-スラッシュコマンドでは、`/np`を使用してください。
+スラッシュコマンドでは、`/nowplaying`を使用してください。
 
 ## 別名
 `現在再生中`以外にも以下の別名を使用できます。
 
 - 今の曲
-- np
 - nowplaying
+- np
