+++ docs/docs/guide/commands/ping.md
@@ -1,15 +1,16 @@
 ---
-sidebar_label: アップタイム
+sidebar_label: ピング
 ---
-# `アップタイム`コマンド
-ボットのアップタイムを表示します。
+# `ピング`コマンド
+ボットのping時間(レイテンシ)を表示します。
 
-スラッシュコマンドでは、`/uptime`を使用してください。
+スラッシュコマンドでは、`/ping`を使用してください。
 
 ## 別名
-`アップタイム`以外にも以下の別名を使用できます。
+`ピング`以外にも以下の別名を使用できます。
 
-- uptime
+- ping
+- latency
