+++ src/Commands/remove.ts
